Reorganised files - Shifted the Structs files into a new folder - Split the Lesson and Timetable Struct to different swift files
