+++ Announcer/Support Files/Globals/Structs/LinksStruct.swift
